internal/registry: remove dead code
(cherry picked from commit 7716219e17d935a6327e244333c1deeddc3574aa)
