+++ internal/registry/config.go
@@ -106,19 +106,6 @@ func newServiceConfig(options ServiceOptions) (*serviceConfig, error) {
 	return config, nil
 }
 
-// copy constructs a new ServiceConfig with a copy of the configuration in config.
-func (config *serviceConfig) copy() *registry.ServiceConfig {
-	ic := make(map[string]*registry.IndexInfo)
-	for key, value := range config.IndexConfigs {
-		ic[key] = value
-	}
-	return &registry.ServiceConfig{
-		InsecureRegistryCIDRs: append([]*registry.NetIPNet(nil), config.InsecureRegistryCIDRs...),
-		IndexConfigs:          ic,
-		Mirrors:               append([]string(nil), config.Mirrors...),
-	}
-}
-
 // loadMirrors loads mirrors to config, after removing duplicates.
 // Returns an error if mirrors contains an invalid mirror.
 func (config *serviceConfig) loadMirrors(mirrors []string) error {
@@ -320,16 +307,10 @@ func ValidateIndexName(val string) (string, error) {
 }
 
 func normalizeIndexName(val string) string {
-	// TODO(thaJeztah): consider normalizing other known options, such as "(https://)registry-1.docker.io", "https://index.docker.io/v1/".
-	// TODO: upstream this to check to reference package
 	if val == "index.docker.io" {
 		return "docker.io"
 	}
 	return val
-}
-
-func hasScheme(reposName string) bool {
-	return strings.Contains(reposName, "://")
 }
 
 func validateHostPort(s string) error {
@@ -354,32 +335,6 @@ func validateHostPort(s string) error {
 		}
 	}
 	return nil
-}
-
-// newIndexInfo returns IndexInfo configuration from indexName
-func newIndexInfo(config *serviceConfig, indexName string) *registry.IndexInfo {
-	indexName = normalizeIndexName(indexName)
-
-	// Return any configured index info, first.
-	if index, ok := config.IndexConfigs[indexName]; ok {
-		return index
-	}
-
-	// Construct a non-configured index info.
-	return &registry.IndexInfo{
-		Name:    indexName,
-		Mirrors: []string{},
-		Secure:  config.isSecureIndex(indexName),
-	}
-}
-
-// GetAuthConfigKey special-cases using the full index address of the official
-// index as the AuthConfig key, and uses the (host)name[:port] for private indexes.
-func GetAuthConfigKey(index *registry.IndexInfo) string {
-	if index.Official {
-		return IndexServer
-	}
-	return index.Name
 }
 
 // ParseRepositoryInfo performs the breakdown of a repository name into a
